pyproject.toml
[tool.poetry]
name = "bd-academia"
version = "0.1.0"
description = ""
authors = ["Artem-Gritsan <gritsanart.ag@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
